Support variable length keys
Previosuly key length was a property of the data file. Records in the data file
stored the key in this fixed number of bytes after the data length marker.

This change removes the limitation of fixed length keys. Data records now
store the key length for each record in two bytes immediately after the data
length marker.